+++ CHANGELOG.md
@@ -26,3 +26,4 @@
 - 搜索结果树支持高亮匹配的关键字 | Search results tree supports highlighting matching keywords.
 - 单击搜索结果在页面左侧预览块 | Click the search result to preview the block on the left side of the page.
 - 配置浏览器扩展开发环境 | Configure browser extension development environment.
+- 搜索块类型支持 `嵌入块` | Search block type supports `Embed Block`.